Replaced in-component function with service, implemented unsubscribe
